Enlever les commentaires et remplacement des int du temps pour des doubles

+++ NexusSimulatHeure/src/Metier/Profil/ProfilPassagerFini.java
@@ -16,7 +16,7 @@ public class ProfilPassagerFini extends ProfilPassager{
     
     private int nombreMax;
     
-    public ProfilPassagerFini(int nombreMax, Point point, int heureDepart, int frequence, DistributionTriangulaire distribution){
+    public ProfilPassagerFini(int nombreMax, Point point, double heureDepart, double frequence, DistributionTriangulaire distribution){
         super(point, heureDepart, frequence, distribution);
         this.nombreMax = nombreMax;
     }
